Moc group creation

+++ cloud/services/groups/groups.go
@@ -26,6 +26,11 @@ import (
 	"k8s.io/klog/v2"
 )
 
+const (
+	TagKeyClusterGroup = "createdBy"
+	TagValClusterGroup = "cloud-operator"
+)
+
 // Spec specification for group
 type Spec struct {
 	Name     string
@@ -49,6 +54,7 @@ func (s *Service) Get(ctx context.Context, spec interface{}) (interface{}, error
 
 // Reconcile gets/creates/updates a group.
 func (s *Service) Reconcile(ctx context.Context, spec interface{}) error {
+
 	groupSpec, ok := spec.(*Spec)
 	if !ok {
 		return errors.New("Invalid group specification")
@@ -58,37 +64,57 @@ func (s *Service) Reconcile(ctx context.Context, spec interface{}) error {
 		// group already exists, cannot update since its immutable
 		return nil
 	}
+	//adding tag to group
+	tag := make(map[string]*string, 1)
+	cloudop := TagValClusterGroup
+	tag[TagKeyClusterGroup] = &cloudop
 
-	klog.V(2).Infof("creating group %s in location %s", groupSpec.Name, groupSpec.Location)
+	klog.V(2).Infof(" DEBUG creating group %s in location %s with tag %v", groupSpec.Name, groupSpec.Location, tag)
 	_, err := s.Client.CreateOrUpdate(ctx, groupSpec.Location, groupSpec.Name,
 		&cloud.Group{
 			Name:     &groupSpec.Name,
 			Location: &groupSpec.Location,
+			Tags:     tag,
 		})
 	if err != nil {
 		return err
 	}
 
-	klog.V(2).Infof("successfully created group %s", groupSpec.Name)
+	klog.V(2).Infof("DEBUG successfully created group %s", groupSpec.Name)
 	return err
 }
 
-// Delete deletes a group.
+// Delete deletes a group if group is created by cloud operator
 func (s *Service) Delete(ctx context.Context, spec interface{}) error {
+	klog.V(2).Infof("DEBUG Deleting group")
 	groupSpec, ok := spec.(*Spec)
 	if !ok {
 		return errors.New("Invalid group specification")
 	}
-	klog.V(2).Infof("deleting group %s in location %s", groupSpec.Name, groupSpec.Location)
-	err := s.Client.Delete(ctx, groupSpec.Location, groupSpec.Name)
+	group, err := s.Client.Get(ctx, groupSpec.Location, groupSpec.Name)
 	if err != nil && azurestackhci.ResourceNotFound(err) {
-		// already deleted
+		// ignoring the NotFound error, since it might be already deleted
+		klog.V(2).Infof("DEBUG group %s not found in location %s", groupSpec.Name, groupSpec.Location)
 		return nil
+	} else if err != nil {
+		return err
 	}
-	if err != nil {
-		return errors.Wrapf(err, "failed to delete group %s", groupSpec.Name)
+	groupObj := (*group)[0]
+	klog.V(2).Infof("DEBUG Got group spec", "Group Spec", groupObj)
+	value, ok := groupObj.Tags[TagKeyClusterGroup]
+	// delete only if created by cloud operator
+	if ok && (value != nil && *value == TagValClusterGroup) {
+		err := s.Client.Delete(ctx, groupSpec.Location, groupSpec.Name)
+		if err != nil && azurestackhci.ResourceNotFound(err) {
+			// already deleted
+			return nil
+		}
+		if err != nil {
+			return errors.Wrapf(err, "failed to delete group %s", groupSpec.Name)
+		}
+		klog.V(2).Infof("DEBUG successfully deleted group %s", groupSpec.Name)
+	} else {
+		klog.V(2).Infof("DEBUG skipping group %s deletion, since it is not created by cloud operator", groupSpec.Name)
 	}
-
-	klog.V(2).Infof("successfully deleted group %s", groupSpec.Name)
 	return err
 }
